cosas machitas hechas, mejoras para el categorias y tambien el popup para que funcione con los no logueados, tambien en las ordenes y tambien en las notificaciones page

+++ src/presentation/components/ofertas/ProductPopup.jsx
@@ -1,16 +1,21 @@
 import React, { useState, useEffect } from 'react';
 import './productpopup.css';
 import { getProductCategoryById } from '../../../infraestructure/api/product_category';
-import {auth,db} from "../../../infraestructure/firebase--config.js";
-import {addDoc, collection, deleteDoc, doc, getDoc, getDocs, query, updateDoc, where} from "firebase/firestore";
+import { auth, db } from "../../../infraestructure/firebase--config.js";
+import { addDoc, collection, deleteDoc, getDocs, query, updateDoc, where } from "firebase/firestore";
+import { useNavigate } from 'react-router-dom';
+import { ToastContainer, toast } from 'react-toastify';
+import 'react-toastify/dist/ReactToastify.css';
 
 const ProductPopup = ({ product, onClose, addToCart }) => {
-    const user = auth.currentUser
+    const user = auth.currentUser;
     const [categoryName, setCategoryName] = useState('');
     const [rating, setRating] = useState(null);
     const [hover, setHover] = useState(null);
     const [isInWishlist, setIsInWishlist] = useState(false);
-    const [selectedRating, setSelectedRating] = useState(null); // Nuevo estado para rastrear la estrella seleccionada previamente
+    const [selectedRating, setSelectedRating] = useState(null);
+    const navigate = useNavigate();
+
     useEffect(() => {
         const fetchCategoryName = async () => {
             try {
@@ -27,17 +32,19 @@ const ProductPopup = ({ product, onClose, addToCart }) => {
     useEffect(() => {
         const checkUserRating = async () => {
             try {
-                const querySnapshot = await getDocs(query(collection(db, 'products_ratings'),
-                    where('product_id', '==', product.id),
-                    where('user_id', '==', user.uid)));
+                if (user) {
+                    const querySnapshot = await getDocs(query(collection(db, 'products_ratings'),
+                        where('product_id', '==', product.id),
+                        where('user_id', '==', user.uid)));
 
-                if (!querySnapshot.empty) {
-                    const userRating = querySnapshot.docs[0].data().rating;
-                    console.log('User rating data:', userRating);
-                    setRating(userRating); // Almacena la calificación del usuario en el estado
-                    setSelectedRating(userRating);
-                } else {
-                    console.log('No user rating found');
+                    if (!querySnapshot.empty) {
+                        const userRating = querySnapshot.docs[0].data().rating;
+                        console.log('User rating data:', userRating);
+                        setRating(userRating);
+                        setSelectedRating(userRating);
+                    } else {
+                        console.log('No user rating found');
+                    }
                 }
             } catch (error) {
                 console.error('Error fetching user rating:', error);
@@ -46,53 +53,38 @@ const ProductPopup = ({ product, onClose, addToCart }) => {
 
         checkIfInWishlist();
         checkUserRating();
-    }, [product.id, user.uid])
-
+    }, [product.id, user]);
 
     const checkIfInWishlist = async () => {
         try {
-            const wishlistRef = collection(db, 'wishlist');
-            const querySnapshot = await getDocs(query(wishlistRef, where('product_id', '==', product.id), where('user_id', '==', user.uid)));
-            setIsInWishlist(!querySnapshot.empty); // Actualiza el estado isInWishlist en función de si el documento existe en la colección de wishlist
+            if (user) {
+                const wishlistRef = collection(db, 'wishlist');
+                const querySnapshot = await getDocs(query(wishlistRef, where('product_id', '==', product.id), where('user_id', '==', user.uid)));
+                setIsInWishlist(!querySnapshot.empty);
+            }
         } catch (error) {
             console.error('Error checking wishlist:', error);
         }
     };
-    /*const handleMouseMove = (e) => {
-        const stars = e.target.parentNode.querySelectorAll('.fa-star');
-        const mouseX = e.clientX - stars[0].getBoundingClientRect().left;
-        const starWidth = stars[0].getBoundingClientRect().width;
-        const score = Math.ceil(mouseX / starWidth);
-
-        const scoreSpan = e.target.parentNode.nextElementSibling;
-        scoreSpan.textContent = `(${score})`;
-
-        for (let i = 0; i < stars.length; i++) {
-            if (i < score) {
-                stars[i].classList.add('fas');
-            } else {
-                stars[i].classList.remove('fas');
-            }
-        }
-    };*/
 
     const addRating = async (productId, rating) => {
         try {
-            // Obtener una referencia a la colección de calificaciones
+            if (!user) {
+                toast.error('Solo los usuarios registrados pueden puntuar. Por favor, inicia sesión.');
+                navigate('/login');
+                return;
+            }
             const ratingsRef = collection(db, 'products_ratings');
-            // Intentar encontrar un documento de calificación existente
             const ratingsQuery = query(ratingsRef, where('product_id', '==', productId), where('user_id', '==', user.uid));
             const querySnapshot = await getDocs(ratingsQuery);
 
             if (!querySnapshot.empty) {
-                // Documento encontrado, actualizar la calificación
                 const ratingDoc = querySnapshot.docs[0];
                 await updateDoc(ratingDoc.ref, {
                     rating: rating,
                     date: new Date().toLocaleDateString()
                 });
             } else {
-                // No se encontró el documento, crear uno nuevo
                 await addDoc(ratingsRef, {
                     product_id: productId,
                     user_id: user.uid,
@@ -101,30 +93,31 @@ const ProductPopup = ({ product, onClose, addToCart }) => {
                 });
             }
             console.log('Calificación actualizada con éxito.');
-            setSelectedRating(rating); // Actualiza el estado local
+            setSelectedRating(rating);
         } catch (error) {
             console.error('Error al actualizar la calificación:', error);
         }
     };
 
-
     const handleWishlistButtonClick = async () => {
+        if (!user) {
+            toast.error('Solo los usuarios registrados pueden agregar a la lista de deseos. Por favor, inicia sesión.');
+            navigate('/login');
+            return;
+        }
         try {
             const wishlistRef = collection(db, 'wishlist');
             if (isInWishlist) {
-                // Si el producto está en la lista de deseos, eliminarlo
                 const querySnapshot = await getDocs(query(wishlistRef, where('product_id', '==', product.id), where('user_id', '==', user.uid)));
                 querySnapshot.forEach(async (doc) => {
                     await deleteDoc(doc.ref);
                 });
             } else {
-                // Si el producto no está en la lista de deseos, agregarlo
                 await addDoc(wishlistRef, {
                     product_id: product.id,
                     user_id: user.uid
                 });
             }
-            // Actualizar el estado isInWishlist después de agregar o eliminar el producto
             setIsInWishlist(!isInWishlist);
         } catch (error) {
             console.error('Error adding/removing product to/from wishlist:', error);
@@ -132,9 +125,7 @@ const ProductPopup = ({ product, onClose, addToCart }) => {
     };
 
     const handleRatingChange = async (currentRating) => {
-        // Verificar si la estrella seleccionada previamente es la misma que la actual
         if (selectedRating === currentRating) {
-            // Si es la misma estrella, no permitir que el usuario vote nuevamente la misma calificación
             return;
         }
         setRating(currentRating);
@@ -201,6 +192,7 @@ const ProductPopup = ({ product, onClose, addToCart }) => {
                     </div>
                 </div>
             </div>
+            <ToastContainer />
         </div>
     );
 };
